Class based rewrite with separate extractors for every site
